Extract box computing logic from matplotlib
